Reformed file system so it can search for files in  base folders, depending on whether a portable or installed build is desired. Also, the audio decoders are moved away from stdio.h file routines and use the asset file system now.

+++ array_macros.h
@@ -3,7 +3,7 @@
 #include <cstddef>
 
 #define ARRAY_COUNT(a) \
-    ((sizeof(a) / sizeof(*(a))) / static_cast<std::size_t>(!(sizeof(a) % sizeof(*(a)))))
+    static_cast<int>((sizeof(a) / sizeof(*(a))) / static_cast<std::size_t>(!(sizeof(a) % sizeof(*(a)))))
 
 #define FOR_N(index, n) \
     for (auto (index) = 0; (index) < (n); ++(index))
